+++ CHANGELOG.md
@@ -860,6 +860,7 @@ This extension assumes a set of default exclude folders:
 - Fixed the detection of not used types from `export { type TYPE }`
 - Fixed the detection of not used types from `export type { TYPE }`
 - Fixed the detection of paths aliases usage
+- Updated the packages
 
 ### Removed